Pull ParameterList children as ParameterItem component
