+++ packages/ui-components/src/components/DataTable/editing/ParameterList.js
@@ -6,14 +6,7 @@
 import PropTypes from 'prop-types';
 import React, { useEffect, useRef, useState } from 'react';
 import styled from 'styled-components';
-import { Divider as BaseDivider } from '@material-ui/core';
-import Grid from '@material-ui/core/Grid';
-import BaseDeleteOutlinedIcon from '@material-ui/icons/DeleteOutlined';
-import { Checkbox as BaseCheckbox, Select, TextField } from '../../Inputs';
-import { IconButton as BaseIconButton } from '../../IconButton';
-import { FlexStart } from '../../Layout';
-import { ParametersType } from './types';
-import { FilterTypeOptions } from '../PreviewFilters';
+
 import { TextButton as BaseTextButton } from '../../Button';
 
 const TextButton = styled(BaseTextButton)`
@@ -27,29 +20,7 @@ const TextButton = styled(BaseTextButton)`
   }
 `;
 
-const Divider = styled(BaseDivider)`
-  margin-bottom: 20px;
-`;
-
-const Checkbox = styled(BaseCheckbox)`
-  .MuiSvgIcon-root {
-    font-size: 0.9rem;
-  }
-  .MuiTypography-body1 {
-    font-size: 0.9rem;
-  }
-`;
-
-const IconButton = styled(BaseIconButton)`
-  top: 35px;
-  width: 30%;
-  height: 20px;
-`;
-const DeleteOutlinedIcon = styled(BaseDeleteOutlinedIcon)`
-  font-size: 25px;
-`;
-
-export const ParameterList = ({ parameters, onDelete, onChange, onAdd }) => {
+export const ParameterList = ({ children, onAdd }) => {
   const parameterListRef = useRef();
   const [scrollToTheBottom, setScrollToTheBottom] = useState(0);
 
@@ -65,88 +36,7 @@ export const ParameterList = ({ parameters, onDelete, onChange, onAdd }) => {
   return (
     <div>
       <div>
-        {parameters.map(
-          ({
-            id,
-            name = '',
-            type = '',
-            required = false,
-            hasDefaultValue = false,
-            defaultValue = '',
-            hasError = false,
-            error = '',
-          }) => {
-            return (
-              <Grid container spacing={0} key={id}>
-                <Grid container spacing={2}>
-                  <Grid item xs={5}>
-                    <TextField
-                      error={hasError}
-                      helperText={hasError && error}
-                      value={name}
-                      placeholder="Text"
-                      label="Name"
-                      onChange={event => {
-                        onChange(id, 'name', event.target.value);
-                      }}
-                    />
-                  </Grid>
-                  <Grid item xs={4}>
-                    <Select
-                      id={id}
-                      value={type}
-                      label="Type"
-                      options={FilterTypeOptions}
-                      onChange={event => {
-                        onChange(id, 'type', event.target.value);
-                      }}
-                    />
-                  </Grid>
-                  <Grid item xs={1}>
-                    <IconButton variant="text" onClick={() => onDelete(id)}>
-                      <DeleteOutlinedIcon />
-                    </IconButton>
-                  </Grid>
-                </Grid>
-                <Grid item xs={12}>
-                  <FlexStart>
-                    <Checkbox
-                      label="Required"
-                      color="primary"
-                      checked={required}
-                      onChange={event => {
-                        onChange(id, 'required', event.target.checked);
-                      }}
-                    />
-                    <Checkbox
-                      label="Add default value"
-                      color="primary"
-                      checked={hasDefaultValue}
-                      onChange={event => {
-                        onChange(id, 'hasDefaultValue', event.target.checked);
-                      }}
-                    />
-                  </FlexStart>
-                </Grid>
-                {hasDefaultValue && (
-                  <Grid item xs={5}>
-                    <TextField
-                      value={defaultValue}
-                      placeholder="Text"
-                      label="Default value"
-                      onChange={event => {
-                        onChange(id, 'defaultValue', event.target.value);
-                      }}
-                    />
-                  </Grid>
-                )}
-                <Grid item xs={10}>
-                  <Divider />
-                </Grid>
-              </Grid>
-            );
-          },
-        )}
+        {children}
         <div ref={parameterListRef} />
       </div>
       <TextButton
@@ -163,8 +53,6 @@ export const ParameterList = ({ parameters, onDelete, onChange, onAdd }) => {
 };
 
 ParameterList.propTypes = {
-  onDelete: PropTypes.func.isRequired,
-  onChange: PropTypes.func.isRequired,
+  children: PropTypes.node.isRequired,
   onAdd: PropTypes.func.isRequired,
-  parameters: ParametersType.isRequired,
 };
