+++ packages/ui-components/src/components/DataTable/editing/index.js
@@ -4,4 +4,5 @@
  */
 
 export * from './ParameterList';
+export * from './ParameterItem';
 export * from './types';
